+++ recipes/config.rb
@@ -13,28 +13,27 @@ rules_location = '/etc/prometheus/prometheus.rules.yml'
 data_location = '/var/lib/prometheus'
 skip_configure = false
 
-scrape_override = node['rblx_prometheus']['config']['telegraf_input']['target_list_override']
+scrape_override = node['rblx_prometheus']['config']['scrape_input']['target_list_override']
 alert_override = node['rblx_prometheus']['config']['alertmanager']['target_list_override']
 prom_override = node['rblx_prometheus']['config']['prometheus']['target_list_override']
 remote_write_override = node['rblx_prometheus']['config']['remote_write']['target_list_override']
 
+##### copy-paste blob to get pod and dc
+infradb_attr = node['infradb'].to_h
+infradb_available = infradb_attr.dig('serverInfo', 'Server') ? true : false
+location_available = infradb_attr.dig('serverInfo', 'Server', 'Location') ? true : false
 
-infradb_available = (node.key?('infradb') and node['infradb'].key?('serverInfo') and node['infradb']['serverInfo'].key?('Server') and not node['infradb']['serverInfo']['Server'].nil?)
-location_available = (infradb_available && !node['infradb']['serverInfo']['Server']['Location'].nil?)
+pod = config['pod'] || infradb_attr.dig('serverInfo', 'Server', 'Location', 'Pod', 'Name')
+raise 'target_discovery: pod unspecified by attributes and graphql unavailable' unless pod || config['pop_mode'] 
+pod = pod.downcase() if pod
 
-pod = node['rblx_prometheus']['config']['pod']
-if pod.nil?
-  raise "#{recipe_name}: pod unspecified by attributes and graphql unavailable" unless location_available
-  pod = node['infradb']['serverInfo']['Server']['Location']['Pod']['Name'].downcase()
-end
+dc = config['datacenter'] || infradb_attr.dig('serverInfo', 'Server', 'Location', 'DataCenter', 'Abbreviation')
+raise 'target_discovery: datacenter unspecified by attributes and graphql unavailable' if dc.nil?
+dc = dc.downcase()
+##### end of copy-paste blob
 
-dc = node['rblx_prometheus']['config']['datacenter']
-if dc.nil?
-  raise "#{recipe_name}: datacenter unspecified by attributes and graphql unavailable" unless location_available
-  dc = node['infradb']['serverInfo']['Server']['Location']['DataCenter']['Abbreviation'].downcase()
-end
-
-req_tags = [pod, dc]
+req_tags = [dc]
+req_tags << pod if pod
 
 alert_list = []
 prom_list = []
@@ -52,13 +51,17 @@ rescue => e
   skip_configure = true
 end
 
-scrape_list = scrape_override.empty? ? node['rblx_prometheus']['config']['telegraf_input']['_target_list'] : scrape_override
+if scrape_override.empty?
+  scrape_hash = node['rblx_prometheus']['config']['scrape_input']['_target_hash']
+else
+  scrape_hash = { dc => scrape_override }
+end
 alert_list = alert_override.empty? ? alert_list : alert_override
 prom_list = prom_override.empty? ? prom_list : prom_override
 remote_write_list = remote_write_override.empty? ? remote_write_list : remote_write_override
 
-log 'debug scrape_list' do
-  message "Final scrape_list is #{scrape_list}"
+log 'debug scrape_hash' do
+  message "Final scrape_hash is #{scrape_hash}"
   level :debug
 end
 
@@ -123,7 +126,7 @@ log 'debug skip_configure' do
 end
 
 log 'debug target_list disable' do
-  message "value of  target_list disable immediately before writing config is #{node['rblx_prometheus']['config']['telegraf_input']['_disable']}"
+  message "value of  target_list disable immediately before writing config is #{node['rblx_prometheus']['config']['scrape_input']['_disable']}"
   level :debug
 end
 
@@ -131,7 +134,7 @@ template config_location do
   source 'prometheus/prometheus.yml.erb'
   variables(
     alert_list: alert_list,
-    scrape_list: scrape_list,
+    scrape_hash: scrape_hash,
     prom_list: prom_list,
     remote_write_list: remote_write_list,
     rr_drop_metric_regex: config['remote_write']['drop_metric_regex'],
@@ -143,5 +146,5 @@ template config_location do
   group 'prometheus'
   notifies :reload, "service[prometheus]", :delayed
   not_if { skip_configure && (alert_list.empty? || prom_list.empty? || remote_write_list.empty?) }
-  not_if { node['rblx_prometheus']['config']['telegraf_input']['_disable'] && scrape_override.empty? }
+  not_if { node['rblx_prometheus']['config']['scrape_input']['_disable'] && scrape_override.empty? }
 end